+++ plutarch-test/tests/Plutarch/Extra/ByteStringSpec.hs
@@ -2,7 +2,7 @@ module Plutarch.Extra.ByteStringSpec (spec) where
 
 import Data.Char (ord)
 
-import Plutarch.Extra.ByteString (pallBS, pisHexDigit)
+import Plutarch.Extra.ByteString (pisHexDigit)
 
 import Plutarch.Prelude
 
@@ -14,9 +14,9 @@ import Test.Hspec
 spec :: Spec
 spec = describe "extra.bytestring" . pgoldenSpec $ do
   "allandhexdigit" @\ do
-    "allas" @| pallBS # plam (#== pconstant (toInteger $ ord 'a')) # pconstant "aaaaaaaaaaaaa" @-> passert
-    "not all as" @| pallBS # plam (#== pconstant (toInteger $ ord 'a')) # pconstant "aaaaaabaaaaa" @-> passertNot
-    "allhex" @| pallBS # pisHexDigit # pconstant "5a7c18eae8778d15344f" @-> passert
-    "notallhex" @| pallBS # pisHexDigit # pconstant "5a7c18eae8778g15344f" @-> passertNot
+    --    "allas" @| pallBS # plam (#== pconstant (toInteger $ ord 'a')) # pconstant "aaaaaaaaaaaaa" @-> passert
+    --    "not all as" @| pallBS # plam (#== pconstant (toInteger $ ord 'a')) # pconstant "aaaaaabaaaaa" @-> passertNot
+    --    "allhex" @| pallBS # pisHexDigit # pconstant "5a7c18eae8778d15344f" @-> passert
+    --    "notallhex" @| pallBS # pisHexDigit # pconstant "5a7c18eae8778g15344f" @-> passertNot
     "pisHexDigit" @| pall @PList # pisHexDigit # integerList (toInteger . ord <$> "1234567890abcdef") @-> passert
     "pisNoneHexDigit" @| pany @PList # pisHexDigit # integerList (toInteger . ord <$> "ghikjklmnopqrstuvwxyz !@#$%^&*()[]{}`~") @-> passertNot